+++ test/interactive.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+
+echo "Hello Program"
+echo "I hope you can process this"
+echo "Please enter your name: "
+
+read varname
+echo "Nice to meet you $varname"
+
+
+printf "And now what is your age: "
+read varage
+
+echo "Ah $varname, you are really $varage?"